Add files via upload
Оптимизация скорости соединения и чтения/записи >64k

+++ firmware/isp.h
@@ -61,12 +61,12 @@ uchar ispEnterProgrammingMode();
 uchar ispReadEEPROM(unsigned int address);
 
 /* write byte to flash at given address */
-uchar ispWriteFlash(uint32_t address, uchar data, uchar pollmode);
+uchar ispWriteFlash(uint16_t address, uchar data, uchar pollmode);
 
-uchar ispFlushPage(uint32_t address);
+uchar ispFlushPage(uint16_t address);
 
 /* read byte from flash at given address */
-uchar ispReadFlash(uint32_t address);
+uchar ispReadFlash(uint16_t address);
 
 /* write byte to eeprom at given address */
 uchar ispWriteEEPROM(unsigned int address, uchar data);
